lai-django/pyproject.toml
[tool.poetry]
name = "lai-django"
version = "0.1.0"
description = ""
authors = ["TheShepsMT <theshepsmt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
openai = "^1.55.0"
django-environ = "^0.11.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
